+++ hardpy/pytest_hardpy/pytest_wrapper.py
@@ -25,7 +25,7 @@ class PyTestWrapper:
         self.config = ConfigManager().get_config()
         self.collect(is_clear_database=True)
 
-    def start(self) -> bool:
+    def start(self, start_args: dict | None = None) -> bool:
         """Start pytest subprocess.
 
         Returns:
@@ -37,44 +37,35 @@ class PyTestWrapper:
         if self.is_running():
             return False
 
+        cmd = [
+            self.python_executable,
+            "-m",
+            "pytest",
+            "--hardpy-db-url",
+            self.config.database.connection_url(),
+            "--hardpy-tests-name",
+            self.config.tests_name,
+            "--sc-address",
+            self.config.stand_cloud.address,
+        ]
+        if self.config.stand_cloud.connection_only:
+            cmd.append("--sc-connection-only")
+        cmd.append("--hardpy-pt")
+        if start_args:
+            for key, value in start_args.items():
+                arg_str = f"{key}={value}"
+                cmd.extend(["--hardpy-start-arg", arg_str])
+
+        if system() == "Windows":
+            self._proc = subprocess.Popen(  # noqa: S603
+                cmd,
+                cwd=ConfigManager().get_tests_path(),
+                creationflags=subprocess.CREATE_NEW_PROCESS_GROUP,
+            )
         if system() == "Linux":
             self._proc = subprocess.Popen(  # noqa: S603
-                [
-                    self.python_executable,
-                    "-m",
-                    "pytest",
-                    "--hardpy-db-url",
-                    self.config.database.connection_url(),
-                    "--hardpy-tests-name",
-                    self.config.tests_name,
-                    "--sc-address",
-                    self.config.stand_cloud.address,
-                    "--sc-connection-only"
-                    if self.config.stand_cloud.connection_only
-                    else "",
-                    "--hardpy-pt",
-                ],
+                cmd,
                 cwd=ConfigManager().get_tests_path(),
-            )
-        elif system() == "Windows":
-            self._proc = subprocess.Popen(  # noqa: S603
-                [
-                    self.python_executable,
-                    "-m",
-                    "pytest",
-                    "--hardpy-db-url",
-                    self.config.database.connection_url(),
-                    "--hardpy-tests-name",
-                    self.config.tests_name,
-                    "--sc-address",
-                    self.config.stand_cloud.address,
-                    "--sc-connection-only"
-                    if self.config.stand_cloud.connection_only
-                    else "",
-                    "--hardpy-pt",
-                ],
-                cwd=ConfigManager().get_tests_path(),
-                creationflags=subprocess.CREATE_NEW_PROCESS_GROUP,  # type: ignore
             )
 
         return True