+++ hardpy/pytest_hardpy/utils/__init__.py
@@ -2,7 +2,13 @@
 # GNU General Public License v3.0 (see LICENSE or https://www.gnu.org/licenses/gpl-3.0.txt)
 
 from hardpy.pytest_hardpy.utils.connection_data import ConnectionData
-from hardpy.pytest_hardpy.utils.const import TestStatus
+from hardpy.pytest_hardpy.utils.const import (
+    ChartType,
+    ComparisonOperation,
+    Group,
+    MeasurementType,
+    TestStatus,
+)
 from hardpy.pytest_hardpy.utils.dialog_box import (
     BaseWidget,
     CheckboxWidget,
@@ -16,11 +22,7 @@ from hardpy.pytest_hardpy.utils.dialog_box import (
     TextInputWidget,
 )
 from hardpy.pytest_hardpy.utils.exception import (
-    DuplicatePartNumberError,
-    DuplicateSerialNumberError,
-    DuplicateTestStandLocationError,
-    DuplicateTestStandNameError,
-    DuplicateTestStandNumberError,
+    DuplicateParameterError,
     ImageError,
     TestStandNumberError,
     WidgetInfoError,
@@ -29,27 +31,39 @@ from hardpy.pytest_hardpy.utils.machineid import machine_id
 from hardpy.pytest_hardpy.utils.node_info import NodeInfo
 from hardpy.pytest_hardpy.utils.progress_calculator import ProgressCalculator
 from hardpy.pytest_hardpy.utils.singleton import SingletonMeta
+from hardpy.pytest_hardpy.utils.stand_type import (
+    Chart,
+    Instrument,
+    NumericMeasurement,
+    StringMeasurement,
+    SubUnit,
+)
 
 __all__ = [
     "BaseWidget",
+    "Chart",
+    "ChartType",
     "CheckboxWidget",
+    "ComparisonOperation",
     "ConnectionData",
     "DialogBox",
-    "DuplicatePartNumberError",
-    "DuplicateSerialNumberError",
-    "DuplicateTestStandLocationError",
-    "DuplicateTestStandNameError",
-    "DuplicateTestStandNumberError",
+    "DuplicateParameterError",
+    "Group",
     "HTMLComponent",
     "ImageComponent",
     "ImageError",
+    "Instrument",
+    "MeasurementType",
     "MultistepWidget",
     "NodeInfo",
     "NumericInputWidget",
+    "NumericMeasurement",
     "ProgressCalculator",
     "RadiobuttonWidget",
     "SingletonMeta",
     "StepWidget",
+    "StringMeasurement",
+    "SubUnit",
     "TestStandNumberError",
     "TestStatus",
     "TextInputWidget",